+++ app/src/main/java/eu/darken/capod/main/ui/overview/cards/SinglePodsCard.kt
@@ -43,6 +43,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import eu.darken.capod.R
+import eu.darken.capod.main.ui.overview.cards.components.AncModeSelector
+import eu.darken.capod.main.ui.overview.cards.components.DebugSection
+import eu.darken.capod.main.ui.overview.cards.components.DeviceConnectionBadge
+import eu.darken.capod.main.ui.overview.cards.components.SignalIndicator
+import eu.darken.capod.main.ui.overview.cards.components.StatusChip
 import eu.darken.capod.common.SystemTimeSource
 import eu.darken.capod.common.compose.Preview2
 import eu.darken.capod.common.compose.PreviewWrapper
