+++ app/src/main/java/eu/darken/capod/main/ui/overview/cards/components/DebugSection.kt
@@ -1,0 +1,38 @@
+package eu.darken.capod.main.ui.overview.cards.components
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import eu.darken.capod.common.compose.Preview2
+import eu.darken.capod.common.compose.PreviewWrapper
+
+@Composable
+fun DebugSection(
+    rawDataHex: List<String>,
+    modifier: Modifier = Modifier,
+) {
+    Column(modifier = modifier) {
+        Spacer(modifier = Modifier.height(8.dp))
+        Text(
+            text = "--- Debug ---",
+            style = MaterialTheme.typography.labelSmall,
+            color = MaterialTheme.colorScheme.onSurfaceVariant,
+        )
+        Text(
+            text = rawDataHex.joinToString("\n"),
+            style = MaterialTheme.typography.labelSmall,
+            color = MaterialTheme.colorScheme.onSurfaceVariant,
+        )
+    }
+}
+
+@Preview2
+@Composable
+private fun DebugSectionPreview() = PreviewWrapper {
+    DebugSection(rawDataHex = listOf("07 19 01 0E 20 75 AA B5 31 00 00 30 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00"))
+}
